Build new fallback auth encrpytors
In new versions of authlogic, the `transition_from_restful_auth` is not
longer going to be available.  This means there is no way to get into
the code branch that adds the AUTH_SITE_KEY to the tokens before hashing
the password for comparison with the db.

Thankfully, they still support a `transition_from_crypto_providers` as a
nice way to update your encryption without having all your users
locked out.

To get around the first issue (`transition_from_restful_auth`) this PR
provides 2 new restful auth encryptors that have custom matchers which
know how the tokens should be structured to make a match.  They should
be used *only* in that transition phase so that eventually everyone
would be moved to the new scheme (SCrypt) in this case.