Remove references to Ruby Sass library (as it is deprecated)
To address the ruby sass deprecation
(http://sass.logdown.com/posts/7081811), we need to remove references
to that gem.

This includes removing the `sass_color_extractor` gem and bumping
`sassc` and `sassc-rails` gems.

In addition, because I needed to rebuild the colors page, I did a bit of
css color variable renaming and cleanup.