Allow admin artist table to be sorted by date times
Problem
-------

With the tabs on the artists admin lists, we were not initializing the
angular app *after* the ajax load.  This mean that any kind of
`<timeago>` angular element would not get initialized and that would get
in the way of sorting by those columns.

Solution
--------

Tell angular not to bootstrap on that page and call the bootstrap
manually after we've pulled all the ajax data.