Update sample data os participation step
Problem
--------

A whlies back (PR #4) we removed the `update_os_participation` method
when we moved to actually joining `Artist` with `OpenStudiosEvent` (as
it should be).  The sample data script was still depending on that
method.

Solution
--------

Update the sample data method to build the join records with a shovel
(`<<`) and not try to use a non-existing method.