Fix sporadic ruby crashes during Rails startup
Problem
-------

Recently I've been seeing hard ruby crashes during the startup of the
app in development.  After some research, it appears there is an issue
with `getaddrinfo` on MacOS and this is causing problems.  Though `ruby`
processes crashed hard, if you were patient, the server would eventually
start up.  But the crashes seemed really weird.

Solution
--------

After reading https://bugs.ruby-lang.org/issues/15763, I found a fix
which was to force unicorn to bind to 127.0.0.1.

This PR adds a unicorn config for development to specify that binding
and removes the unused 'puma' config and 'puma' gem.