Handle the case when there are no open studios events
Problem
-------

The `remind_for_open_studios_register?` method is counting on
`OpenStudiosEventService.current` returning a thing.  But it's possible
that there are no events to return (rare, but possible).

Solution
--------

Return false to remind folks if there is no current open studios event.
And... if there is one, do the rest of the logic.