Add the ability to promote (or not promote) an open studios event
Problem
-------

Given the coronavirus, we actually don't want to promote this particular
event.  Promotion means that we'd encourage artists to sign up for open
studios when we're within 12 weeks of the event.  In this case, we
actually want to turn off that promotion since we're trying to take
people off the list.

Solution
--------

Add a `promote` column to the open studios event (default true) and then
if that column is turned off, the reminder in the nav will not show up.