Add site preferences model
We want to save Social Media tags which are to be used in some system
emails.  It made sense to add a `SitePreferences` model which can just
hold whatever "global" things we might have to save.  Currently it'll
just be social media tags.

Then we can use it in the new_art_piece_presenter to send the new custom
tags to the folks watching for new email.