Replace `browser.modern?` with use of `browserslist-useragent`
The `browser` gem has been reporting that `modern?` will be deprecated
in an upcoming release and should be replaced with some other scheme.

This PR follows the pattern outlined here
https://dev.to/amplifr/outdated-browser-detection-with-browserslist-10co
taking advantage of `browserslist-useragent` to determine whether or not
the browser is supported or not.

Changes
-------
* Add a script to update `browsers.json`
* update our code to use that list for the new `supported_browser?`
method
* update `old_browser_warning` to be a little prettier
* add `.browserlistrc` which we will use for both autoprefixer and
generating the `browsers.json`