Harden tests using the selectize helpers
Problem
-------

The selectize helpers that we have had trouble when the `value` that we
were looking for in dropdowns contained an apostrophe because it busted
the interpolated script string that we use to test selectize.

Solution
--------

Revamp the way we quote the value when we make the javascript command
used to select from the dropdown.