Move to webpack for all the things
Problem
-------

We have no chance of using modern javascript without webpack.  So let's
try to move everything (and i mean *everything*) in webpack.  This
includes images, fonts, stylesheets and javascript.

Solution
--------

Move all asset compilation to webpacker.

This includes images, fonts, stylesheets and javascript.

Changes
-------
* cleanup around admin styles
* Move back to yarn from `npm ci`
* Improve fa icons used and fa icon helpers
* Move to using imports
* normalize $.ajax calls to a `mau_ajax` library
* move MAU.Utils and MAU.Flash into their own libraries and import them
as needed
* remove teaspoon in favor of jest (still some angular tests that need updating)
* Move easy js specs to jest
* Move to imported spin.js function
* revamp hookup to datatables
* move QueryStringParser to a lib/module
* Better handling of promote checkbox and testing around it
* .gitignore ignores tgz files
* Fix ngInject issue with favoriteThis
* Move to inhouse JS Google Maps handling
* update circle so cucumber does not precompile assets but depends on
the cache (see note below)

Notes:

Moving forward let's try to move all the JS to ES6 (var => let/const
etc) and leverage all the power we have with ES6 to make the code look a
little better.

Over time, the JS will likely get moved around a bit.  The current
directory structure is weird and could be condensed probably.

Finally, the circle config now is using the cache fairly heavily.
Because running `webpacker:compile` and `elasticsearch` in the same run
causes Elasticsearch to crash, the `prepare` step now does *not* start
up ES (it never really needed to) and now does the `assets:precompile`.
The cucumber step just restores that cache so if there are failures then
we're in trouble.  But this is because `assets:precompile` will try to
compile even if it doesn't need to.  So calling the `compile_assets`
step is not sufficient to not make webpack build assets twice.