Fix Cookie Serializer to support old sessions with new cookies
Problem
-------

With the recent update to Rails6, if you had a session cookie, it
was not being properly read by the system because it assumed different
serialization scheme (the old Rails 5 version).

Solution
--------

Force the app to do a hybrid approach so it can handle both the old and
the new cookies.