Fix icons for search (after font awesome upgrade)
Problem
-------

After our font awesome upgrade (which came with webpacker upgrade),
we lost the icons in the search

We need to update the fa-* classes so we get the right icons.

Solution
--------

Update icons we choose for the search hit and update scss to accomodate
both `fa` and `far` icons for color.