Add some hooks to streamline the webpack build
Problem
-------

webpack build takes up a pile of memory and is maybe too much for little
vm's like on circleCi or on our deployed apps vms where ElasticSearch is
trying to use a bunch of resources.

Solution
--------

Attempt to trim the resources required to build webpack by using
`filesystem` instead of `memory` for caching.

Changes
-------
* split custom config into separate items so we can add as needed
* add `assets:clobber` task to rake
* add .nvmrc