Revamp Elasticsearch indexing and querying to clear up deprecations
Problem
-------

We get a huge pile of deprecation warnings from ES because of the way
we have the ngram min/max stuff setup.

Solution
--------

Revamp the way we index so we're more inline for ES 7 and stop getting
deprecation warnings (and maybe do a little better on the way we index
and search).

Indexes will need to be rebuilt after this goes out.

Changes
-------

* dump `_all` field in favor of simpler indexing
* add `english` analyzer for art piece title
* update ngram params to avoid deprecation warnings