Javascript overhaul for better testing
Problem
-------

We kind of need a better JS organization so we can actually test things
properly

Solution
--------

The main move here is to try to move the `jQuery( init function )` kind
of logic to instead build new classes (or jquery plugins) which are then
added to the `MAU` global namespace of functions.  Then the
initialization of that plugin can be done in the slim/html template so
that it's much more obvious when things get initialized and how the js
is tied to the dom.

In doing this, we provide access to those classes or plugins which make
them much easier to test.

Also in this PR, we fixed the issues with the `htmlLoader` which was preventing
good testing of angular components in jest. The solution is to rewrite a
dummy loader just for html which is basically cribbed from
[`webpack-contrib/raw-loader`](https://github.com/webpack-contrib/raw-loader/blob/master/src/index.js).

Changes
-------

* Fix html loader thing for jest
* migrate `editable_content` spec
* dump `spec/javascripts`
* Add a tests for notify_mau
* Add test for SearchHit model
* Add medium.js component test
* add simple test for jquery.feedback component
* Add responseTransformer $resource helper and cleanup around that
* add tests to share button then cleanup and refactor
* add tests for notification service
* add test for repeater delimiter
* Trim admin and application.js to (hopefully) include only what they need)
* Add test for search service
* Move files to `.service.js` from `_service.js`
* Add test for appplication event service and looking to using humps
* Add more unit tests for ngResource dependent services
* Init ApplicationEventsDetailsForm in the view

Problem
-------

The code we use for the application events query form is hard to test
because it's just a pile of jQuery which get's bound to the DOM when we
load that Javascript.

Solution
--------

Instead - so we could imagine testing it better, make the code into a
class which can be exported.  Then provide it to the app as a global.
Finally, only do the binding of the JS to the DOM in the template where
we render that piece of DOM.

Changes
-------
* move application event details form code to a class so it can be exported
* add tests for that component
* add `admin/globals` where we import anything we need global and stick
it on `window`
* add the jQuery binding to the template (instead of in the js)

* Move artists admin index initializers to globals and init in the view
* Add tests for markdown and cleanout markdown from admin/tests
* Remove unused hash_change_observer
* Move mau_spinner to mau/spinner and test spinner
* Move to camel case variables in query string parser
* Move account type chooser to global and test it
* Extract `artists.js` into plugins
* Extract main.js in to separate plugins/classes
* Move whats this popup to init  in view template
* Move sampler to template init
* move paypal form to its own template initialized plugin
* Extract AccordionShowSectionGivenHash from users.js and dump that file
* trim navigation.js and bust into nav and mobile_nav components
* Move art arrangement js to a component
* Move admin/studios to studio_arranger component
* Move admin/roles to admin_role_manager.js
* Revamp admin datatables hookup to use components
* eslint and move admin graphs to components