Precompile assets locally and rsync for deployment
Problem
-------

Deployments with capistrano were recompiling assets on the deployed
server which was taking up resources that the server did not have and
causing ElasticSearch to fall over during deployment or causing the
asset compilation to fail because there wasn't enough memory.

Solution
--------

Precompile assets on the deploy*ing* machine and then sync them to the
deployed machine.