Revamp handling of media/tag frequency
Problem
-------

We had a pile of code wrapped around computing and caching the
frequency of media and tag usage so we could render those pages
with things in order of use.  but the code was not very efficient.

Solution
--------

Leverage SQL group by to compute frequency of both tags and media
which should be more efficient and allows us to forego the caching
of that stuff.

Changes
-------
* compute frequency of media or tag use with smart sql
* align media cloud presenter and tag cloud presenter code
* add media service and trim art piece tag service
* update controllers etc to handle these new trim services