Remove mode support for art piece tag and media pages
Problem
-------

we had a bunch of code to support showing media/art_tags by artist or
art piece.  This mode was supported with our old UI but the new UI does
not need this switch.

Solution
--------

Clean out the code that supports that mode in favor of only showing
tags/media by art piece.

Changes
--------

* Clean up paginators and presenters to not support mode
* Add per_page initializers in the config so they are easier to change
* Remove unused `SearchQuery` object