Performance Memory Tuning
Problem
-------

After noticing that `bundle` is taking 2G on our deployed boxes
and digging in with `derailed` (and other reading), it appears
that the `append_view_path` is causing memory issues.  My suspicion
is that the path continues to get appended to with each request
making a global array that grows for the lifetime of the app.

Solution
--------

Don't use append view path at all and just specify the partials
with a full path.

Additionally, we were doing a lot of extra file reading to check
for the supported browser, so I put that in a service and memo-ized
what I could.