Move to jsonapi-rb for serializer
Problem
--------

Active model serializers has run it's course.  It is preventing some
other bundle updates.  So it's time to move.

After some research it seems the `jsonapi-rb` is a fast and `jsonapi`
compliant solution.  Not that we really *need* the `jsonapi` format but
it is a standard and will hopefully be pretty stable going forward.

Solution
--------

Replace `api/v2` endpoints that we use in angular with the new
`jsonapi-rb` serializers.   Then rebuild the api connection in angular
to use this new shape and build models that can gracefully consume the
shape and hand it to the frontend.

Changes
-------
* convert serializers to use `jsonapi` base serialization.
* Add a new `json_api` module in js that just consumes our server json
  api endpoints
* move away from angular `$resource` because we can't easily get in
  there and wrap responses with our models.  Also because as we move to
  React, we may want a more generic `api` solution that is not dependent
  on our JS framework.
* backfill old json returning endpoints to include the root key (which
  was previously provided by AMS) - see
  `admin/application_events_controller` and `admin/emails_controller`

Notes
-----

In doing this work I noticed we're using `jquery.ajax` for all our ajax
calls.  We might think about moving that to use `fetch` some day so we
are a little less tied to `jQuery`