Add a `Mailer` React component to replace angular `mailer` directive
Problem
--------

We're moving to React!  Let's choose a very simple component and figure
out how to move forward with React.

Solution
--------

Add a Mailer react component that does what the angular component does.
Then revamp the way we mount react components in the app to make it a
little easier to add stuff as we go.  Finally move the react test
package to a styleguide that includes (at this point) all known react
components with examples of how to use them.

Changes
-------
* move from `mount` to an `automount` javascript system that will scrape
  the page for all components with a known class (`react-component`) and
  render the react component in place
* update `react_component` helper to require `id` and `component` to be
  passed in to allow multiple versions of the component on the same page
  without conflicting dom id parameters
* add tests for the `react_component` helper
* add `react_styleguide` page which contains some documentation of
  current available components and code samples
* use the new `Mailer` component on the signup page
* add new `stylesheets/gto/components` directory and `index.scss` file
  to contain component styles
* add `componentRegistry` which should hold a list of all components
  that need to be imported for Rails to render.  Any internal components
  (imported only by other React components) need to be in this list.
* add jest and react-testing-library for testing
* add react development info to the readme