Fix angular loading issue
Problem
--------

With the recent moves around angular code, we were not properly
ng-annotating which meant for production builds, where we minimized and
uglified things, dependency injection was no longer working.

We'd included the proper Babel package to handle this for us but we'd
neglected to turn it on.

Solution
--------

Turn on `babel-ng-annotate`

Changes
-------
* include `angularjs-annotate` in babel config
* yarn upgrade all the easy upgrade
* clear out a console.log