Studios are mis tagged with OS start
Problem
--------

We were not showing open studios stars on Studio cards properly.

https://www.pivotaltracker.com/story/show/176730215

Solution
--------

It appears that with a recent efficiency change, we ended up showing
stars on studios if *anyone* was doing open studios instead of if anyone
*in this studio* was doing open studios.

Solved this with a small fix in the StudioPresenter and added a test to
confirm.