Update the way we handle Address on artist
Problem
-------

Something recently changed with delegates making our strange delegation
strategy to get the address for an artist doesn't work as well as it
used to.  Not sure what changed.  I suspect a rails upgrade.  Basically
what was starting to happen was that `artists.lat` was no longer
`artist.studio.lat` in cases where it should have been.

https://www.pivotaltracker.com/story/show/176730160

Solution
--------

Rethink the way we handle address.

Instead of letting the `Address` model figure out how to find the
address stuff, we let the artist decide which model it should get the
address from and let `Address` simply be a decorator/presenter for
either a studio or artist_info - both which contain real address
information.

The rest of this code is to push that change through consuming
components and update specs