Add price to art piece [176835246]
As an artist
I should be able to add "price" to my art piece
and I should see it on the site.

https://www.pivotaltracker.com/story/show/176835246

Solution
--------

Add `price` field on art piece and send it through the API to the
frontend as both `price` (raw) and `display_price` which is currency
formatted.

In the art piece browser (art piece detail page), show the formatted
price if we have it.

In edit and new art piece forms, add the field for price.

Add cucumber tests