Handle bug with editing artist with bad info
Problem
--------

We were blowing up on editing artist if you had validation errors
because we were not getting all the data setup for the view - in the
failure case.  And we had no tests around that.

Additionally, we were doing a bad job of surfacing errors if you edited your artist
info but there were errors in the data (like blank email).  The section
that contained the error would remain closed and there was no flash to
tell you something went wrong.

Solution
--------

If we hit an error during `update` and we decide to `render :edit`, lets
get all the data ready for the view.

Then in the view, we add some Javascript to check for any sections that
have `.error` and if there are any, open that accordion up.

Finally, added a cucumber test to prove it.