Add phone number to user
Problem
-------

We want to be able to surface an artist's phone number (with their
consent) for the upcoming virtual open studios work.  So first we need
to store their phone number some where.

Though on some level, this probably belongs on the ArtistInfo, it also
goes along with basic user data and because the current `basic_info`
form is the same (ish) for users and artists alike, putting phone on
users makes more sense for the form data and easier to get the data
surfaced.

https://www.pivotaltracker.com/story/show/176835330

Solution
--------

Add phone to user and surface it for artists only in the basic_info edit
form.

Changes
-------
* add field
* fix #update and #edit methods for users and artist to better surface
  errors
* update cucumber specs

fix tests