Add validation on art_piece content size. max 4MB
Problem
--------

Uploading big files that lead to web timeouts leads to the error page
which is no good for users.

Solution
---------

For now, we're going to put a 4M limit on it and hope that saves most
folks from trouble.  A better long term solution would be client-side
file uploaders with a progress bar.  Hopefully this will help as a
bandaid until we get to that point.

Changes
-------

* Add content size validation on ArtPiece.

Note
----

We have 2 other image upload opportunities - studio image (which is
admin only so it doesn't really matter) but the artist profile image
also might want this treatment.  Currently it has a slightly different
validation scheme so we should probably roll that out with care which is
why I didn't do that work here.