Add client side file size validation
Problem
-------

We added backend validation with rails models but those are done on the
server which means we need to upload the whole file before we can tell
what size it is which is a bit counter intutitive.

Solution
--------

Add a client side check so we can *immediately* give the user feedback
that they are uploading too big an image.

Changes
--------

* Add clientside validation
* remove notion of a thumbnail as it looks like it's basically unused.
  which might help processing time with uploads
* Remove studios/_thumb partial as it appears unused