Provide an open studios info form for an artist doing open studios
Problem
-------

Artists who are doing open studios with the new virtual studios event
need new information.

This information is in the db (as of #121).

Solution
--------

Revamp the frontend OpenStudiosRegstration component (which was angular) to be in react
and allow for input of the OpenStudiosParticipant information.

Note: Because we're just getting started with React - there are bunch of foundational changes
in this PR like where we put types, and testing and Formik and some other stuff.

Changes
--------

* Add OpenStudios components (OpenStudiosRegstrationSection, OpenStudiosRegistration, OpenStudiosInfoForm)
* Add Formik to support form input
* Replace existing angular component with react version
* add cucumber tests
* change logic around `doing_open_studios?` - use presence of `OpenStudiosParticipant`
* start moving away from supporting OpenStudios lookup by key (which should be no longer required)