Better sample and seed data
Problem
-------

I got this setup on a new computer and had some trouble getting some sample data to play with.

Solution
---------

Fix up seed data to include roles and update sample data to be a bit more forgiving.
I suspect we made a change around the FavoritesSevice.add method and forgot to update the
users sample data file to match that update.  For now, just let it ride.