Remove use of OS keys (and reduce use of OpenStudiosEventShim)
Problem
--------

We used to have a hard coded list of OS events maintained by date keys.

With recent work (#4) we moved to having a proper model connected to
artists to manage this relationship.

`OpenStudiosEventShim` was a tool used to help with backwards
compatibility as we moved through that.

Since then, we've migrated data so that all the previous keys have their
own proper event which is linked to artists based on the previously used
record in "deprecated_open_studios_participation" on the `ArtistInfo`
model.

Solution
--------

In this PR we remove the use of the hard coded keys
(`available_open_studios_keys`) which were provided by the shim and do
what can be done to reduce the use of the shim across the app.

Followup
--------

Eventually we could probably remove the shim all together because now
all it provides is the `current_key` which is
`OpenStudiosEvent.current,key`.  But we can leave that for another PR>

Additionally, the `deprecated_open_studios_participation` field
should/could be removed in a future PR.