remove unused notify_mau react components
Problem
-------

In a recent cleanup #112, this code (which was in process) got merged
in.  Once we changed tack to focus on virtual open studios, this fell by
the way side.  Since that time we've introduced a form component
(`Formik`) which was what this work was spiking on.

Solution
--------

Take out the trash for now.