Add Youtube Video to OpenStudiosParticipant.
Problem
---------

as an artist doing open studios
I can add a video link to my open studios info
so that people can get a studio tour or whatever
even if i'm not present to share that.

Solution
--------

Add `youtube_url` to the OpenStudiosParticipant model and allow users to
update that information in their `OpenStudiosInfoForm`.

Changes
-------

* Add new field on OpenStudiosParticipant model
* Add `YoutubeUrlValidator` to validate that it looks like a youtube link
* Add the field to the OpenStudiosInfoForm
* Move MauHint to use `p` over `div` for proper spacing