Update open studios promo page
Problem
--------

As an visitor or artist
before the event
I can get information about the upcoming event
so i'm excited to participate as an artist or a visitor

In this move, we are making the current `/open_studios` page which used
to be the landing page for the event into more of a promotional page.

When we talk about it well talk about the Open Studios Promo Page which
will be distinct from the Open Studios Market page which will be the event
landing page when that page and the subsequent artist pages underneath
feel more like a separate website.

Solution
--------

Modify the current `/open_studios` page to *only* have a list of participating
artists (without grouping by studio) and some content that is managed (as it was)
by the CMS system.