Add special event time and date to the open studios event model and update os index page looking to that change
problem
-------

ideally we should know if there is a special event timing inside the
*main* event timing.

In service of https://www.pivotaltracker.com/story/show/176835274

solution
--------

Add that.

Now OpenStudiosEvent has a special_event set of date and time fields
which are must be *inside* the main date range.

These are used in the open studios registration text and later may be
used for switching on and off other features/data.

changes
-------
* add new fields for special event time and date range
* Shuffle date range helpers off together in preparation for special event time window helpers
* put that value conditionally in `openStudiosFormInfo` instructions
* add date validations and tests