Add timeslots to open studios event
For the special events, we want to know a list of available
timeslots in some duration (now fixed to 1 hr).

After talking with @rdavid1099 we agreed on a new field that is
auto computed based on the special event start/end date and time.

As you update the event, we can re-record available time slots.

Then consumers (when we want artists to choose from available time
slots) can read that list of timeslots and use it for building forms
or whatever.

This PR is just the datamodel and save changes to support that.

Also we make sure that we compute the timeslots in the timezone where
events will happen (which we'll hardcode to PST cuz that's where we
live and work).