Add some TS lint
Problem
--------

We added typscript but are *far* from obeying the rules.

Solution
---------

Take a half step there by adding a few linters and getting eslint to
checkout our Typescript code and complain as it seems appropriate.

`yarn tsc` still fails with lots of errors - mostly related to not
knowing the `@...` directory aliases.  For now, let that go.

Changes
--------

* add a couple ts related lint plugins
* run the linter and clean up so it passes