Don't auto register me [177278822]
Problem
--------

In the past, we were trying to help artists find the registration flow.
We did this by opening the "Would you like to register" dialog if they
came in from certain links.

With the new registration flow, this feels like it's more confusing than
it is helpful.  Because we *now* have an open studios registration form
(instead of just a yes/no toggle) we have visual indicators that you're
registered or not.  So we don't need the auto register anymore.

Solution
--------

Take out the logic that tries to auto open the registration dialog.

Changes
-------

* dump the flash that records if we should auto register
* remove the logic to automatically pop up the dialog
* update cucumber tests
* always show the dialog (if you change your reg status)

Note
----

That dialog may feel weird in terms of language but future stories will
change it to have separate confirm dialogs for when you register or
unregister.