Art piece navigation breaks when studio is indy [177294826]
Problem
-------

The artpiece page breaks if the studio is independent.

With our new json_api serializers, we were not converting null to 0
and the backend was returning a 400 causing the rest of the rendering to
fail.

Solution
---------

Integerize the id from the front so we call for studio/0.json and the
backend knows to return the 'independent studio' record.