Better handling of the promo banner
Problem
-------

The promo banner got rushed out and it didn't have a title or
any seo/accessibiity stuff.

Solution
---------

Take the time to get a title on the link for the promo banner.
Also give the promo thing a little gutter space so it looks like it
fits with all the other images.