ARG TIMEZONES! Ensure special event date/times are in event time zone
Problem
--------

Timezones!  We're sending date strings from the server and not doing any
time zone interpretation.

https://www.pivotaltracker.com/story/show/177331037
https://www.pivotaltracker.com/story/show/177331045

Solution
--------

For special event start and end dates we really care about the timezone
because we use that to compute the timeslots.  Let's make sure we put
them in the right timezone when we start.

Changes
-------

* Coerce incoming dates to pst before we put them in the database
* when editing times, if we can't generate new timeslots, blank them out