Move catalog stuff to `catalog_for_print` to not get confused
Problem
--------

We have a `catalogs` controller and a little view that is intended to
make a catalog list for printing. This is different from our current
notion of the open studios catalog.

Solution
--------

Rename a few files to get them out of the way and make it so we don't
confuse the catalog with the old catalog for print.