Open Studios Catalog page includes artists [176835248]
Problem
--------

As a visitor
when i visit the open studios catalog I should see a custom page
which shows the details of the event
So that I can focus on the most important part of what open studios is about - the artists!

Solution
--------

Update the current main catalog page to include a list of participant
artist pictures and a cms section that can be updated by an admin
to include a little bit of content at the top.

Changes
--------

* share OpenStudios presenter logic
* add new artist/art cards for the open studios catalog
* build the catalog page