Add ability to zoom into an art piece [177238660]
Problem
--------

Now that we have an artists page that shows art, we have no good way to
see the full version of the art or it's info (price, dimensions etc).

Solution
--------

Build that functionality.

To accomplish this we need the art cards to be in React so that when you
click on them we can launch a modal with the full version of the art and
whatever details we want.

For this I built an `ArtModal` component which given an id, fetches the
art and renders the art piece in a closeable window.

Then I built `ArtCard` which is rendered by Rails to wrap the art card
details that sit on the artists show page with the modal so the entire
card *is* the trigger to get the modal to show up.

Incidental adds where the `CloseButton` component and the `Spinner`
component.