Prevent open studios routes from falling back [177334098]
Problem
---------

Currently you can visit openstudios.mau/studios which is *not* supposed
to be a route in the catalog site.

https://www.pivotaltracker.com/story/show/177334098

Solution
---------

Add constraints for *all* routes so there is no fallback.

This is not *quite* true.  We've left `api/*` routes and
`admin/cms_documents/*`to be available on both sides of the house
because new react stuff uses the api endpoints and the we need
cms_documents editing in both places.

Also, add error controller so 404's in the subdomain use the subdomain
layout instead of the main.

Finally support `www` or '' for the main domain.