Trim schedule date format so it fits better - artists' catalog show
Problem
---------

as a visitor
when i hit the artist show page in the catalog
the schedule dates should be formatted as follows

4/23 2-3pm PST
4/24 1-2pm PST

we can get away with this because we know that we are setting the special event timing to be on the hour. so for now... we can get away without the minutes and the year.

the current format is much longer

03/23/2021 2:00pm - 3:00pm PDT
03/24/2021 3:00pm - 4:00pm PDT

https://www.pivotaltracker.com/story/show/177431666

Solution
---------

Build a "compact" version of the time slot formatter and use it.