Build ArtPage in reactjs
Problem
--------

We want to make the art piece page include a browser/carousel.

This means we need to fetch all art pieces.  Since we *also* need that
for the art page, instead of having rails render a collection of art
cards, let's just fetch all the art and hand it to a react component.

Solution
---------

React ArtPage component.  This replaces a large portion of the artist
show page.

wip

Add art page

lint

tests