See the next art piece in the modal
Problem
-------

as a user visiting the catalog
when i visit an artist's studio page
and i click on a piece of art so i see a modal
i should be able to click next
and get the next piece in the modal.

https://www.pivotaltracker.com/story/show/177431680

Solution
--------

Build out carousel functionailty inside the modal.

Changes
-------

* new useCarouselState to provide previous/next/current
* add ArtPiecesContext to hold the list of all art pieces
* consume the context at the modal level and hook up next/previous
  buttons

Note
----

I'm leaving the keyboard triggers to a separate story because they were
suprisingly tricky.