Add current open studios info to the artists csv download
Problem
--------

As event producer I want to make sure all the artists are
ready to go before the event and that we don't send our
visitors to an artist with an open sign that is not
actually open. There will be sign-up errors, changed minds,
and incomplete forms. I want to communicate with the artists
before the event to uncover these.  To make this task feasible,
I'd like a csv file of the current open studios form data.

Solution
---------

Update the `AdminArtistList` presenter to include open studios
participant info if present in the csv