Add sold treatment to the modal art browser.
Problem
---------

we added a nice treatment for sold on the small art card but not in the
art window in the modal.

Solution
--------

Add a sold floating tag on the art window if the art has sold.

Notes
-----

because of the auto contain of the image on that modal, pinning this to
the bottom right corner is hard so i centered it instead.