Open Studios Catalog shows artist stars after the first 40
Problem
--------

That page was set up to pull the first 40 artists, and after that
it falls back to infinite scroll which is pulling from the main site
using the main site design instead of the open studios rendering.

https://www.pivotaltracker.com/story/show/177801595

Solution
--------

Make that page pull the first 100 artists and hope that we don't have
more than 100 for this first run.

If we get up to 100 in the next 2 weeks, we can fix it properly which
would be to make the infinite scroll understand which domain it's asking
about.