Fix flaky test around subdomains
Problem
--------

There was a problem with our subdomain routing tests because we were
setting the Rails config in the test.  If those tests ran before the
server status tests, then the subdomain for the remaining test suite
would be making urls with the subdomain.

https://www.pivotaltracker.com/story/show/177802908

Solution
--------

This fix is to wrap before and after with the setting of the host
so that when the subdomain routing tests are finished, the environment
host remains in tact.