Sort artists in gallery by "who's open now"
Problem
--------

During the event, we'd like to promote those folks who are actively
broadcasting so if they are, they should be preferentially sorted to the
top of the catalog.

All others should be sorted by last name as they are now.

https://www.pivotaltracker.com/story/show/176835301

Solution
---------

Replace the `ArtistsGallery` presenter we were using with a new
`OpenStudiosCatalogArtists` that can take care of the sorting which is
specific to this page.

I also reworked the main controller to not worry about pagination
because we're not doing any (yet) and so it can use a simpler presenter
(not `ArtistsGallery` which has pagination built in).

Smaller changes were to make the boolean flags that the presenters get
be keyword args instead of just params so some tests needed to be
modified.